Pull

+++ src/user/delete.js
@@ -3,7 +3,6 @@
 var async = require('async'),
 	db = require('../database'),
 	posts = require('../posts'),
-	user = require('../user'),
 	topics = require('../topics'),
 	groups = require('../groups'),
 	plugins = require('../plugins'),
@@ -12,6 +11,9 @@ var async = require('async'),
 module.exports = function(User) {
 
 	User.delete = function(uid, callback) {
+		if (!parseInt(uid, 10)) {
+			return callback(new Error('[[error:invalid-uid]]'));
+		}
 		async.waterfall([
 			function(next) {
 				deletePosts(uid, next);
@@ -40,7 +42,7 @@ module.exports = function(User) {
 	}
 
 	User.deleteAccount = function(uid, callback) {
-		user.getUserFields(uid, ['username', 'userslug', 'email'], function(err, userData) {
+		User.getUserFields(uid, ['username', 'userslug', 'fullname', 'email'], function(err, userData) {
 			if (err)  {
 				return callback(err);
 			}
@@ -53,7 +55,14 @@ module.exports = function(User) {
 					db.deleteObjectField('userslug:uid', userData.userslug, next);
 				},
 				function(next) {
-					db.deleteObjectField('email:uid', userData.email.toLowerCase(), next);
+					db.deleteObjectField('fullname:uid', userData.fullname, next);
+				},
+				function(next) {
+					if (userData.email) {
+						db.deleteObjectField('email:uid', userData.email.toLowerCase(), next);
+					} else {
+						next();
+					}
 				},
 				function(next) {
 					db.sortedSetsRemove([
@@ -67,13 +76,16 @@ module.exports = function(User) {
 				function(next) {
 					var keys = [
 						'uid:' + uid + ':notifications:read', 'uid:' + uid + ':notifications:unread',
-						'uid:' + uid + ':favourites', 'user:' + uid + ':settings',
+						'uid:' + uid + ':favourites', 'uid:' + uid + ':followed_tids', 'user:' + uid + ':settings',
 						'uid:' + uid + ':topics', 'uid:' + uid + ':posts',
 						'uid:' + uid + ':chats', 'uid:' + uid + ':chats:unread',
-						'uid:' + uid + ':ip', 'uid:' + uid + ':upvote', 'uid:' + uid + ':downvote',
+						'uid:' + uid + ':upvote', 'uid:' + uid + ':downvote',
 						'uid:' + uid + ':ignored:cids'
 					];
 					db.deleteAll(keys, next);
+				},
+				function(next) {
+					deleteUserIps(uid, next);
 				},
 				function(next) {
 					deleteUserFromFollowers(uid, next);
@@ -100,6 +112,23 @@ module.exports = function(User) {
 			});
 		});
 	};
+
+	function deleteUserIps(uid, callback) {
+		db.getSortedSetRange('uid:' + uid + ':ip', 0, -1, function(err, ips) {
+			if (err) {
+				return callback(err);
+			}
+
+			async.each(ips, function(ip, next) {
+				db.sortedSetRemove('ip:' + ip + ':uid', uid, next);
+			}, function(err) {
+				if (err) {
+					return callback(err);
+				}
+				db.delete('uid:' + uid + ':ip', callback);
+			});
+		});
+	}
 
 	function deleteUserFromFollowers(uid, callback) {
 		db.getSetMembers('followers:' + uid, function(err, uids) {
